requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.5.0
asgiref==3.8.1
bidict==0.23.1
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colorlog==6.8.2
distro==1.9.0
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
fastapi==0.115.0
filelock==3.16.1
fsspec==2024.9.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.0
idna==3.10
jiter==0.5.0
kombu==5.4.2
openai==1.46.1
packaging==24.1
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.4
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
serpy==0.3.1
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.38.5
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
wsproto==1.2.0
